Fix Windows EC2 onboarding when using a secondary AWS account.
When onboarding a Windows instance from a secondary AWS account, the
Elasticity lambda is configured to use an AssumeRole called
CyberArk-AOB-AssumeRoleForElasticityLambda in order to acquire the
necessary permissions.
A typo in instance_processing::get_instance_password_data() prevents
the session token from being retrieved, and consequently to onboard
the instance. This commit fixes the issue.